New process (wuthout memory initialization) implemented.
Scheduler should be renamed to runner. The runtime is actually the scheduler.
The responsibility of the runner is simply to run code until the list of
code (process list now) is empty. The runtime takes all decisions on how to
manipulate the current list. Each time a code execution yields, the runner
delegates on the runtime to see what happens (debug, new process, end...) and
the scheduler takes the actual action (call debug, add process, delete
current...).