Split expressions in subexpression on calls.
When calling a function or creating a new process, DIV allows the process that
has been called to be paused and inspected. For more than simple calls, this
means to be able of inspecting each call step by step:

result = a(b()) + c();

That expression should be debuggable at:

result = a(b()) + c();
         ^ ^      ^    Debug points
         2 1      3    Order

For Div2js state machines, that means to yield on every call. Something like:

1:
   return __yieldCallFunction(2, b, []);
2:
   return __yieldCallFunction(3, a, [<result of the call at 1>]);
3:
   return __yieldCallFunction(4, c, []);
4:
   mem[...] = <result of the call at 2> + <result of the call at 3>;
   return __yieldEnd;

Thus, returning to the scheduler, the execution can be intercepted.

The procedure to translate the expression syntax tree into its subexpression
form is conceptually simple. Define what is a subexpression and, starting with
the root node:

  1. Translate the sub trees.
  2. Translate current node. Make `output` be the translated node.
  3. If current node is not a subexpression, return `output`.
  4. If current node is a subexpression:
     1. Get an expression for a unique lvalue (i.e. something able to store
        data). Make `aux` to be that expression.
     2. Emit `output` in such a way its result is stored in the lvalue referred
        by `aux`.
     3. Return `aux`

For the former expression, given only calls are subexpressions, the execution
is something like:

                  a
                 /
          2: call         b
         /       \       /
        /         3: call
    1: +                 \
        \         c       ()
         \       /
          4: call
                 \
                  ()

First subexpression is 3. We allocate aux_b and emit 3 like this:

    var aux_b = b();

Now we ascend to 2. We allocate aux_a and emit 2 like:

    var aux_a = a(aux_b);

We reach 4, now. We allocate aux_c and emit 4 like:

    var aux_c = c();

Finally we are in step 2 of node 1. Left node has returned `aux_a` and right
node has returned `aux_c`, the resulting + node is:

    aux_a + aux_c;

Since the + node is not considered a subexpression, we have finished.

But the problem with this approach is that it does not yield so interception
is impossible. But in order to yield we need to return from the JS object code
with an actual `return` statement which destroys current function frame and so
all the local intermediate values.

To solve this problem, a non-local shared memory space is needed for the
scheduler to store the result of the callee and the caller to retrieve the
returned value.

Several solutions are possible. Those solutions must take into account
recursive calls. I chose a return values queue. The scheduler will queue the
result of the callee and the caller will dequeue the value. So, with current
solution, the output code is something like:

1:
   return __yieldCallFunction(2, b, []);
   // scheduler will do returnValues.queue(<result of b()>)
2:
   return __yieldCallFunction(3, a, [returnValues.dequeue()]);
   // scheduler will do returnValues.queue(<result of a(...)>)
3:
   return __yieldCallFunction(4, c, []);
   // scheduler will do returnValues.queue(<result of c()>)
4:
   mem[...] = returnValues.dequeue() + returnValues.dequeue();
   return __yieldEnd;

This is not very readable but it fulfill all our requirements right now.

I'm worried about the solution could bite me if inspecting the same expression
multiple times due to the execution of .dequeue() several times.