Adding global memory browser.
For the render to be able of reading memory so it can render the frame, a memory
browser, similar to that used during debugging is needed. For this reason, we
provide a memory browser as part of the runtime.

Since the memory browser needs the memory map, and the render is part of the
system, this questions if we need a separated two differents compiling modes.

  1. Yes. This would provide DEBUG and NORMAL. Debug will provide the memory
  map for general debug sessions while normall will prevent the map for being
  included in the compiled program. If not included, the render can not use
  the memory browser to read the processes' values so it must rely on constant
  values.

  2. No. This option leaves the memory map in the compiled program and allows
  general debugging sessions and rendering without coupling the values.

If runtime is combined with the compiled AST program, a posterior static code
analysis can transform option 2 into 1 and it leads the responsibility out of
div2js.

(NOTE: To do this, we should annotate constant values so they can be statically
replaced. Consider switching to ES6 as output).