Handling privates.
Translation tests and more in-depth execution tests are required.

In this context, the main program count as another process.

Privates are defined per process. When determining the context in div2checker.js, it is necessary to add the declared privates into the symbol map.

The symbol map in memory/symbols.js has now special methods for adding a new process' private variable and determining if an identifier is a private for a given process.

The memory mapper in memory/mapper.js has been updated to take into account the maximum private segment of all the processes when determining the size of the process for allocating the process tool. In DIV, the process size is always the size of the local segment + the max. size of the private segments.

There are now additional segments, once per process name.

The memory browser has been modified to accept the type of the process so we can find the proper privates in the correct process but this is artificial. All process types should be mapped to a number stored in a local variable per process. Since local initialization is still not supported, an explicit indicator of the type of the process is needed but this is just a temporal workaround.

The translation module div2trans.js is now capable of translating privates. To do so, it starts by declaring the private offset inside a process which should equal the DIV's local segment size.

When translating the main program or a process, the translator instruct the context in context.js to enter a process scope in which the privates would be valid.

Private variables are translated into JS vars with the offsets each variable. They are added right before the infinite loop used for the body's state machine. Opposite to locals and globals, private names are in lower case. The schema is available in templates.js