Decreasing coupling.
The checker at div2checker.js is in charge of performing type checking and extracting semantics.

The checker is no longer directly dependent of symbols. Also, now the checker can try to extract semantics from partial ASTs.

The symbols module at memory/symbols.js no longer loads the definitions on its own. It now exposes a SymbolTable class which accepts a definitions object from which builds the symbol table.

As a result, the compiler at compiler.js needs to import both the symbol table constructor and the DIV2 definitions to create the proper symbol table and feed it to the checker to extract the proper context from the DIV2 AST.

The translation test has changed to use the new architecture and also removes the fake contexts of the sample DIV ASTs.

Since the context is no longer faked but actually extracted from the real AST, this will lead to a partial rewriting of those tests involving private variables or processes.

This commit also removes an out of date dependency from the AST translator to context.