Adapting translation tests.
The test suite trranslation.spec.js does not fake the context. It is left to the checker to extract a correct context. But it fakes the definition table for well known locals and globals.

It led to a rewriting of most of the tests since now it is not possible to test specific parts of the nodes to translate. Unit tests here would be more useful.

The tests related to arguments have been disabled. This is something we should consider sooner than later.