Allow the scheduler to block on some promise.
DIV2 executed in a synchronous environment (MS-DOS) where calls
to the network or filesystem were bloking. However, given the
inherently asynchronous nature of the Web, implementation of such
calls can no longer be done in a blocking way.

So, in order to emulate this blocking, the scheduler allow to
freeze all process execution until a promise is resolved. Calling
such a method will assume the blocking call happens inside of the
current execution context and, once its done, it will continue
resuming the aforementioned context.