The new tests are in place for refactoring the functions available at runtime. In particular, these two new tests check load_pal() and load_pal() failing when the palette does not exist.
The latter has induced the addition of relevant functionality
related to blocking functions. In particular, now the scheduler
detects when a blocking function fails, and its able of delegate
handling the error.

The runtime could then recover (not done right now) from the
failure and resume the execution.

The scheduler has been simplified to be more stratightforward to
follow. Still, I need to document it.