Remove interfacing with JS from the scheduler responsibilities.
By the way of defining an interface between native JavaScript code
and DIV processes, it is now clear to me that the scheduler should
know about DIV processes only.

JavaScript code is needed to implement the built-in functionality
of DIV. Given the asynchronous nature of JavaScript, chances are
that DIV process need to wait for asynchronous code to finish.

Prior to this commit, the scheduler dealt with it. Now, the
runtime is in charge of interfacing with native JavaScript code
although it is likely it will end encapsulated in some foreign
function interface (FFI) component.

Same native code can fail. In that case, it is still the
responsibility of the runtime to deal with the error.