Several refactors: builtins, hooks, and types for the scheduler.
Here is a risky commit with several changes. My bad.

First, it extracts the built-in functions to its own module and
specifies a simple JavaScript/TypeScript interface:

  - Native functions are passed the set of systems as last parameter.
  - When using synchronous JavaScript, return a number.
  - When using asynchronous JavaScript, return a promise resolving
    into a number.

In addition to the yield baton, now the scheduler also passes the
yielding process.

Second, it reduces the hooks for the scheduler to the following:

  - onyield, for when a process has yield control.
  - onupdate, for when all process have finished running.
  - onfinished, for when the process list becomes empty.

Additionally, now the runtime has a resume() function to resume the
scheduler avoiding the need of exporting a control object to resume
the execution of processes.

Finally, the commit add types to the Scheduler. The scheduler is now
generic in the process type, and provides the minimum interface for
it to work.