Expose Runtime interface for better testability.
Here is the rationale behind. The process map receives the DIV runtime
as a parameter so all the translated DIV process can access runtime
functionality. And so, the runtime functionallity should be publicly
accessible (or, at least, publicly known).

However, due to the asynchronous nature of JavaScript in comparison to
the blocking behaviour of DIV, control cannot be simply handled over
runtime and viceversa. We need a component reconciling sync and async
code.

This component is the scheduler. The scheduler is in charge of running
the synchronous code of the processes until frame completion. When a process
wants to communicate with the runtime, it returns a message to the scheduler
(a Baton object), and the scheduler invokes an "onyield" callback passing
the message, and the originator process.

Previous to this commit, the Baton object was handled by a private
function called "_schedule", now conviniently renamed to "_handle" so it
hints the function is some sort of message handler.

The message handler used to identify a private function with the desired
runtime behaviour. After the commit, that functionality is now public,
although the message handler remains private and ends up calling some of
the public interfaces.

In my mind, the process should be able to call the runtime functionallity
directly. However, that requires adapting all runtime interfaces to return
promises.

This is enough so far.