Implement transparent flag for xput().
Since it is not possible to put a color in the screen that is not
in the palette, making a graph transparent consist in calculating
the blend of color, then finding the closest color in the palette
according to the euclidean distance between color components.

The color blending I suppose DIV is doing is the 50% mix between
whatever was in the screen and the graph color. I don't get 100%
equal results for the following test:

```
PROGRAM xput_example;

PRIVATE
  file1;
  x_coord;
  y_coord;
  angle1;
  size1;
  flags1;

BEGIN
  file1 = load_fpg("help/help.fpg");
  LOOP
    x_coord = 160;//rand(0, 319);
    y_coord = 100;//rand(0, 199);
    angle1 = 0;//rand(-pi, pi);
    size1 = 100;//rand(10, 200);
    flags1 = 0;//rand(0, 7);

    // Put graphic 101
    xput(file1, 101, x_coord, y_coord, angle1, 100, 0, 0);
    xput(file1, 101, x_coord, y_coord, angle1, 200, 4, 0);

    LOOP
    	FRAME;
    END
  END
END
```

In DIV emulation in Boxer (DOS-BOX), the center of the triangle
is brighter. In the web browser, not only the center is dimmer but
there is also a lighter column on the left of the center.

I hope to catch it during the unit testing stage, but this is
not a big deal. I'm satisfied enough.