Draft drawing the processes.
Now systems, when running, receive a reference to the whole runtime,
which allows for improved delegation.

For the changes needed to draw the processes, we need to expose the
process alive at a frame. This is now possible trough the runtime
scheduler.

Finally, the foreground layer where the process are drawn needs for
clearing every frame. DIV2 allows for a partial cleaning of layer
regions but this is not supported just yet. Right now, the cleanse
is complete and, at the beginning of each frame, the process layer
is completely wiped out.