Draft the input system for mouse control.
The input system listens to the mouse events on the canvas and
update the corresponding memory in DIV2.

The performance is not superb but that's less of a concern now.

Next steps should be a review of the architecture of:

- The loader and the way of registering systems.
- The runtime and the order in which running the systems.
- The video system, nomenclature, its public interface, and its modularization.

Also, we should try to type and modernize the translation module.

And finally write some fine grained tests for the new code.