Fix string parsing.
String parsing was done through the use of the JSON.parse() function
which parse string values according to the rules of JavaScript which
are not the rules of DIV2. So, for instance, the path
"tutorial\tutor0.fpg" was parsed as "tutorial    utor0", interpreting
the sequence "\t" as a tab.

This commit fix the parsing.