Fix the defaault directory finder.
My belief was that DIV tried to find a resource in its default directory
when no other directory was specified in the path. For instance,
"DIV.PAL" would become "PAL/DIV.PAL", whereas "HELP/DIV.PAL" would remain
"HELP/DIV.PAL".

This is not true. After some file system experimentation, moving folders
around, what happens is that DIV always tries the default dir, and if
the path file does not exist there, then it tries from the working dir.

This commit modifies the behaviour of the file system to try fetching
from within the default dir, and if this fails, then it tries from within
the rootURL.