Implement declaration of private variables in parameters.
Passing parameters in DIV2 is a bit special with respect to how
other languages treat parameters.

While in other languages, all parameters are local to the function
they are bound to, in DIV, the names in the parameter list can
refer to global, local, and private data declared in the GLOBAL
and LOCAL sections of the main program, or in the PRIVATE section of
the process.

Moreover, all names not in any of these sections are implicitely
added to the process PRIVATE section, as if they were declared there
in the first place, with the same order they appear in the parameter
list.

Another oddity of the DIV2 language is a parameter list can repeat
parameters and so, they will be initialized more than once.