Centralize all declarations within the context.
Declaring new names like those of the program, process, or new
globals, locals, or privates should be managed by the semantics
layer. The ultimate artifact output of this stage is the context
object. The context is needed for the actual translation into
JavaScript.

The commit centralizes all declaration operations within the
context object. Thus, one of the questions about if process names
should be symbols is no longer a problem, since that's part of how
the context keep track of names.

The rules for new names have been reviewed. Names are checked in
strict order of aparition. Following DIV grammar rules, this means:

  - Name of the PROGRAM.
  - Names in the GLOBAL section.
  - Names in the LOCAL section.
  - Names in the PRIVATE section of the program.
  - Name of the first declared PROCESS.
  - Names in the PRIVATE section of the first declared process.
  - Name of the second declrated PROCESS.
  - Names in the PRIVATE seciton of the second declared process.
  - ...

DIV honors the golden rule of "no name can obscure another name".
It translates basically into the following declaration rules:

  1. No name can appear twice in its scope.
  2. No global name can share the name of some private name.

Regarding 1, there are two practical scopes in DIV: global and
private to a process. Names that appear in the CONST, GLOBAL,
and LOCAL sections, and the name of the main program, and those
of the processes are considered of global scope, since they can be
accessed from every code block in the program.

The ones in the PRIVATE section are considered private to the
process in which the section appears.

So, if a name has been declared in the GLOBAL section, the name
cannot appear in the LOCAL sections. But also, because of 2, it
cannot appear in a PRIVATE section.

For instance:

```
PROGRAM main_program;
GLOBAL
  one_variable;
LOCAL
  one_variable; // ERROR `NameIsNotNewError`. `one_variable` is not new.
BEGIN
END

PROCESS one_process();
PRIVATE
  one_variable; // ERROR `NameIsNotNewError`. `one_variable` is not new.
BEGIN
END
```

Now privates can reuse names if they belong to different processes,
but a new global name cannot reuse a name previously declared a
private name since then, also because 2, the private name would obscure
the global name.

For instance:

```
PROGRAM main_program;
BEGIN
END

PROCESS one_process();
PRIVATE
  one_private;
BEGIN
END

PROCESS one_private(); // ERROR `NameIsNotNewError`. `one_private` is not new.
BEGIN
END
```

All these cases are captured in the new semantics.spec test suite. Right now,
the context only takes care of ensuring the rules of declarations but, in the
future, the process of building the context will check other semantics as well.

It is likely a builder pattern is better suited for building the
context. For now, this is fine.