Compute process type values and translate TYPE unary expressions.
Includes initializing the reserved.process_type datum with the
correct value, which has led to modify some tests to include the
initialization of the program per se.

DO NOT continue with this kind of initialization. Not until trying
making the process implementation modify the actual process memory
thorugh its attributes. It should be easy to do reusing a process
view. It should also work with the initialization of the parameters.

This way should the compiled program do not have runtime
responsibilities.