Fix precedence between logical and relational expressions.
The DIV manual says logical operatos have more priority than
relational operators. It cause an expression of the form:

```
y>220 OR collision(TYPE bullet)
```

To be parsed as:

```
   >
  / \
 y   \
     OR
    /  \
  220   \
     collision
         |
        TYPE
         |
       bullet
```

Or, with parenthesis:

```
y>(220 OR collision(TYPE bullet))
```

This expression does not make sense in the context of the
tutorial program TUTOR0B.PRG which evidences the expected
behaviour is for the relational expression to take precedence
over the logical operation. I.e. it should be:

```
(y>220) OR (collision(TYPE bullet))
```

Which corresponds to the following syntax tree:

```
    OR
   /  \
  >    \
 / \    \
y  220   \
          \
       collision
           |
          TYPE
           |
         bullet
```

The changes in this commit change this precedence.

Also, the test altered here ensure the correction. It is expected
the expression inside expression.prg and the one inside
parenthesised-expression.prg translate to the same AST. The
difference is that the expression in parenthesised-expression.prg
has parenthesis in the expected natural priority.

Maybe it should be better to have this, and also another parenthesised
test for making sure we can change the priorirty.