Implement collision() built-in.
This is a fast implementation of the built-in collision(), one of
the key functions of DIV. This implementation does not check
transparent pixels, it simply calculate there is intersection in
the screen bounding boxes of the sprites, and it does not take care
of returning the whole list of colliding processes.

However, implementing collision() encouraged me to try an
entity-component-system approach in which systems can return
specific views of the processes with data relevant to those systems
only.

What I would expect is instead of doing this:

```js
const processView = runtime.getMemoryBrowser().process({
  id: process.id
});
const boundingBox = [
  processView.local("reserved.box_x0"),
  processView.local("reserved.box_y0"),
  processView.local("reserved.box_x1"),
  processView.local("reserved.box_y1"),
];
```

Would be doing something like:

```js
const processGeometry =
  runtime.getComponent(Geometry, process);
const boundingBox = processGeometry.boundingBox;
```

It seems like a good opportunity for trying out the
process/component implementation in which the process memory is
modified directly.