Remove initialized in favor of leveraging the state machine.
Because of the way we translate the body of a process, information
related to initialization is not available during body translation.
So, previous attempts to leverage the state machine of the process
failed due to my inability to add a "0" initialization state.

However, I've realized I can achieve something similar initializing
the programIndex (program counter) to that 0 state and make the
initialization code subject to that value. The result is a simplied
state machine of two states prepended to the body state machine.

In a nutshell, this:

```
if (exec.programIndex === 0) { // initialization state machine
  // Initialization code.
  exec.programIndex = 1;
}
while (true) {                 // process body state machine
  switch (exec.programIndex) {
    case 1:
      // Code for state 1
    ...
  }
}
```