Implement pre and post step events for the scheduler.
The scheduler now triggers listeners before and after the step
invoking onstepstart() and onstepend() handlers. The interface
has changed to allow systems to implement onStepStart() and
onStepEnd() (instead of run()) functions to be executed at the
beginning of the frame (as in WebInputSystem) and at the end
of the frame (as in VideoSystem).

The InputSystem now no longer relies on explicit width and height
for resolution but on the canvas internal resolution.